requirements.txt
streamlit
supabase
pandas
openpyxl
python-dotenv
python-dateutil